+++ src/graphql/timeAndSalary.js
